Реализованы все задания перед упаковкой в Docker
