+++ bot/loader.py
@@ -5,4 +5,5 @@ from dotenv import load_dotenv
 load_dotenv()
 API_TOKEN = os.getenv("BOT_TOKEN")
 bot = Bot(token=API_TOKEN)
-dp = Dispatcher()
+dp = Dispatcher(bot=bot)
+
