+++ calendarapp/api_views.py
@@ -1,0 +1,15 @@
+from rest_framework import viewsets
+from .models import User, Event, Appointment
+from .serializers import UserSerializer, EventSerializer, AppointmentSerializer
+
+class UserViewSet(viewsets.ModelViewSet):
+    queryset = User.objects.all()
+    serializer_class = UserSerializer
+
+class EventViewSet(viewsets.ModelViewSet):
+    queryset = Event.objects.all()
+    serializer_class = EventSerializer
+
+class AppointmentViewSet(viewsets.ModelViewSet):
+    queryset = Appointment.objects.all()
+    serializer_class = AppointmentSerializer
