+++ requirements.txt
@@ -8,6 +8,7 @@ asgiref==3.8.1
 attrs==25.3.0
 certifi==2025.4.26
 Django==5.2.3
+djangorestframework==3.16.0
 frozenlist==1.7.0
 idna==3.10
 magic-filter==1.0.12
